third_party/blink/web_tests/fast/events/touch/send-oncancel-event-expected.txt

Tests that the cancel touch event is sent correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

PASS event.type is "touchcancel"
PASS cancelEvent is non-null.
PASS cancelEvent.pageX is 25
PASS cancelEvent.pageY is 25
PASS successfullyParsed is true

TEST COMPLETE
